// types.ts
export enum LoanStatus {
  EN_ATTENTE = 'EN_ATTENTE',
  APPROUVE = 'APPROUVE',
  DEBLOQUE = 'DEBLOQUE',
  REMBOURSE = 'REMBOURSE',
  REJETE = 'REJETE',
  DEFAUT = 'DEFAUT',
}

export enum InstallmentStatus {
  EN_ATTENTE = 'EN_ATTENTE',
  PAYEE = 'PAYEE',
  EN_RETARD = 'EN_RETARD',
  IMPAYEE = 'IMPAYEE',
  EN_ATTENTE_CONFIRMATION = 'EN_ATTENTE_CONFIRMATION',
}
// ...
export interface Installment {
  installmentNumber: number;
  dueDate: string;
  dueAmount: number;
  paidAmount: number;
  paidAt?: string;
  status: InstallmentStatus;
  daysLate: number;
  warningsSent: boolean;
  paymentProofUrl?: string;
}

export interface LoanApplication {
  _id: string;
  userId: string;
  status: LoanStatus;
  requestedAmount: number;
  fees: number;
  netAmount: number;
  loanPurpose: string;
  validatedBy?: string; // Admin ID
  validationJustification?: string;
  rejectionReason?: string;
  approvedAt?: string;
  disbursedAt?: string;
  completedAt?: string;
  installments: Installment[];
  createdAt: string;
  updatedAt: string;
}